Rename task and file
