+++ crates/services/consensus_module/poa/src/pre_confirmation_signature_service/trigger.rs
@@ -1,7 +1,7 @@
 use super::*;
 use std::future::Future;
 
-/// Defines the behavior for when the `PreConfirmationTask` should rotate the delegate key
+/// Defines the behavior for when the `PreConfirmationSignatureTask` should rotate the delegate key
 pub trait KeyRotationTrigger: Send {
     fn next_rotation(&self) -> impl Future<Output = Result<()>> + Send;
 }
